requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Async HTTP Client
httpx>=0.25.0

# Database & Validation
pydantic>=2.5.0
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# Scheduling (for NotificationsAgent)
apscheduler>=3.10.0

# External APIs
twilio>=8.10.0
openai>=1.0.0

# Date/Time handling
zoneinfo; python_version < "3.9"

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0